add imgui (#50)
* feat(sponge): clean up includes

* feat(imgui): add imgui

* feat(imgui): use imgui 1.89.9

- highest version for non-windows build

* feat(imgui): refactor imgui to layer

* feat(imgui): fix imgui

* fix(build): fix assert

- Because assert is a function-like macro, commas anywhere in the argument that are not protected by parentheses are interpreted as macro argument separators. There is no standardized interface to add an additional message to assert errors. A portable way to include one is to use && with a string literal:

* fix(imgui): remove extraneous code

* feat(imgui): update imgui text

- update copyright

* feat(imgui): move imgui layer to game

* feat(imgui): display imgui on debug build

+++ sponge/src/platform/opengl/openglsprite.cpp
@@ -23,7 +23,7 @@ OpenGLSprite::OpenGLSprite(std::string_view name) : name(name) {
     ebo = std::make_unique<OpenGLElementBuffer>(
         indices, static_cast<uint32_t>(sizeof(indices)));
     ebo->bind();
-    ebo->setData(indices, static_cast<uint32_t>(sizeof(indices)));
+    ebo->setData(indices, sizeof(indices));
 
     uint32_t program = shader->getId();
     if (auto location = glGetAttribLocation(program, "vertex");
@@ -56,7 +56,7 @@ void OpenGLSprite::render(glm::vec2 position, glm::vec2 size) const {
     const auto tex = OpenGLResourceManager::getTexture(name);
     tex->bind();
 
-    vbo->setData(vertices, static_cast<uint32_t>(sizeof(vertices)));
+    vbo->setData(vertices, sizeof(vertices));
 
     glDrawElements(GL_TRIANGLES, static_cast<int32_t>(numIndices),
                    GL_UNSIGNED_INT, nullptr);
